商品上下架

+++ ant-deliver-service/src/main/java/com/bx/ant/controller/ApiItemController.java
@@ -3,9 +3,9 @@ package com.bx.ant.controller;
 import com.bx.ant.pageModel.session.TokenWrap;
 import com.bx.ant.service.ShopItemServiceI;
 import com.bx.ant.service.session.TokenServiceI;
+import com.mobian.absx.F;
 import com.mobian.pageModel.*;
 import com.mobian.pageModel.ShopItem;
-import com.mobian.service.MbItemServiceI;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -37,10 +37,15 @@ public class ApiItemController extends BaseController {
     public Json getAllItemList(HttpServletRequest request,MbItem mbItem, PageHelper ph) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        DataGrid dataGridItem = shopItemService.dataGridWithQuantity(mbItem, ph,token.getShopId());
-        j.setObj(dataGridItem);
-        j.setSuccess(true);
-        j.setMsg("获取成功！");
+        if(!F.empty(token.getShopId())) {
+            DataGrid dataGridItem = shopItemService.dataGridWithQuantity(mbItem, ph, token.getShopId());
+            j.setObj(dataGridItem);
+            j.setSuccess(true);
+            j.setMsg("获取成功！");
+        }else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -57,12 +62,17 @@ public class ApiItemController extends BaseController {
     public Json getShopItemOnline(HttpServletRequest request, ShopItem shopItem, PageHelper ph) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        shopItem.setShopId(token.getShopId());
-        shopItem.setOnline(true);
-        DataGrid dataGridShopItem = shopItemService.dataGridWithItemName(shopItem, ph);
-        j.setObj(dataGridShopItem);
-        j.setSuccess(true);
-        j.setMsg("获取成功！");
+        if(!F.empty(token.getShopId())) {
+            shopItem.setShopId(token.getShopId());
+            shopItem.setOnline(true);
+            DataGrid dataGridShopItem = shopItemService.dataGridWithItemName(shopItem, ph);
+            j.setObj(dataGridShopItem);
+            j.setSuccess(true);
+            j.setMsg("获取成功！");
+        } else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -78,12 +88,17 @@ public class ApiItemController extends BaseController {
     public Json getShopItemOffline(HttpServletRequest request, ShopItem shopItem, PageHelper ph) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        shopItem.setShopId(token.getShopId());
-        shopItem.setOnline(false);
-        DataGrid dataGridShopItem = shopItemService.dataGridWithItemName(shopItem, ph);
-        j.setObj(dataGridShopItem);
-        j.setSuccess(true);
-        j.setMsg("获取成功！");
+        if(!F.empty(token.getShopId())) {
+            shopItem.setShopId(token.getShopId());
+            shopItem.setOnline(false);
+            DataGrid dataGridShopItem = shopItemService.dataGridWithItemName(shopItem, ph);
+            j.setObj(dataGridShopItem);
+            j.setSuccess(true);
+            j.setMsg("获取成功！");
+        } else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -98,9 +113,14 @@ public class ApiItemController extends BaseController {
     public Json updateBatchItemOnline(HttpServletRequest request, String itemIds) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        shopItemService.updateBatchItemOnline(itemIds, token.getShopId());
-        j.setSuccess(true);
-        j.setMsg("批量门店商品上架成功！！");
+        if(!F.empty(token.getShopId())) {
+            shopItemService.updateBatchItemOnline(itemIds, token.getShopId());
+            j.setSuccess(true);
+            j.setMsg("批量门店商品上架成功！！");
+        } else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -115,9 +135,14 @@ public class ApiItemController extends BaseController {
     public Json updateItemOnline(HttpServletRequest request, Integer itemId) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        shopItemService.updateItemOnline(itemId,token.getShopId());
-        j.setSuccess(true);
-        j.setMsg("上架成功！！");
+        if(!F.empty(token.getShopId())) {
+            shopItemService.updateItemOnline(itemId, token.getShopId());
+            j.setSuccess(true);
+            j.setMsg("上架成功！！");
+        } else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -132,25 +157,36 @@ public class ApiItemController extends BaseController {
     public Json updateBatchShopItemOffline(HttpServletRequest request, String itemIds) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        shopItemService.updateBatchShopItemOffline(itemIds, token.getShopId());
-        j.setSuccess(true);
-        j.setMsg("批量下架成功！！");
+        if(!F.empty(token.getShopId())) {
+            shopItemService.updateBatchShopItemOffline(itemIds, token.getShopId());
+            j.setSuccess(true);
+            j.setMsg("批量下架成功！！");
+        } else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
     /**
      * 修改某种商品下架
      *
-     * @param shopItemId
+     * @param itemId
      * @return
      */
     @RequestMapping("/updateItemOffline")
     @ResponseBody
-    public Json updateShopItemOffline(HttpServletRequest request, Integer shopItemId) {
+    public Json updateShopItemOffline(HttpServletRequest request, Integer itemId) {
         Json j = new Json();
-        shopItemService.updateShopItemOffline(shopItemId);
-        j.setSuccess(true);
-        j.setMsg("下架成功！！");
+        TokenWrap token = tokenService.getToken(request);
+        if(!F.empty(token.getShopId())) {
+            shopItemService.updateShopItemOffline(itemId,token.getShopId());
+            j.setSuccess(true);
+            j.setMsg("下架成功！！");
+        }else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -165,25 +201,36 @@ public class ApiItemController extends BaseController {
     public Json deleteBatchShopItem(HttpServletRequest request, String itemIds) {
         Json j = new Json();
         TokenWrap token = tokenService.getToken(request);
-        shopItemService.deleteBatchShopItem(itemIds, token.getShopId());
-        j.setSuccess(true);
-        j.setMsg("批量删除成功！！");
+        if (!F.empty(token.getShopId())) {
+            shopItemService.deleteBatchShopItem(itemIds, token.getShopId());
+            j.setSuccess(true);
+            j.setMsg("批量删除成功！！");
+        }else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
     /**
      * 删除门店某种商品
      *
-     * @param shopItemId
+     * @param itemId
      * @return
      */
     @RequestMapping("/deleteShopItem")
     @ResponseBody
-    public Json deleteShopItem(HttpServletRequest request, Integer shopItemId) {
+    public Json deleteShopItem(HttpServletRequest request, Integer itemId) {
         Json j = new Json();
-        shopItemService.delete(shopItemId);
-        j.setSuccess(true);
-        j.setMsg("删除成功！！");
+        TokenWrap token = tokenService.getToken(request);
+        if(!F.empty(token.getShopId())) {
+            shopItemService.deleteShopItem(itemId,token.getShopId());
+            j.setSuccess(true);
+            j.setMsg("删除成功！！");
+        }else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -191,16 +238,22 @@ public class ApiItemController extends BaseController {
      * 获取门店商品库存量
      *
      * @param request
-     * @param shopItemId
+     * @param itemId
      * @return
      */
     @RequestMapping("/getShopItemQuantity")
     @ResponseBody
-    public Json getShopItemQuantity(HttpServletRequest request, Integer shopItemId) {
+    public Json getShopItemQuantity(HttpServletRequest request, Integer itemId) {
         Json j = new Json();
-        ShopItem shopItem = shopItemService.get(shopItemId);
-        j.setObj(shopItem);
-        j.setSuccess(true);
+        TokenWrap token = tokenService.getToken(request);
+        if(!F.empty(token.getShopId())) {
+            ShopItem shopItem = shopItemService.getByShopIdAndItemId(token.getShopId(),itemId);
+            j.setObj(shopItem);
+            j.setSuccess(true);
+        }else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 
@@ -208,15 +261,22 @@ public class ApiItemController extends BaseController {
      * 修改门店商品库存量
      *
      * @param request
-     * @param shopItemId
+     * @param itemId
      * @return
      */
     @RequestMapping("/updateShopItemQuantity")
     @ResponseBody
-    public Json updateShopItemQuantity(HttpServletRequest request, Integer shopItemId, Integer quantity) {
+    public Json updateShopItemQuantity(HttpServletRequest request, Integer itemId, Integer quantity) {
         Json j = new Json();
-        shopItemService.updateShopItemQuantity(shopItemId, quantity);
-        j.setSuccess(true);
+        TokenWrap token = tokenService.getToken(request);
+        if(!F.empty(token.getShopId())) {
+            shopItemService.updateShopItemQuantity(itemId,token.getShopId(), quantity);
+            j.setSuccess(true);
+            j.setMsg("修改信息成功");
+        }else {
+            j.setSuccess(false);
+            j.setMsg("获取门店信息失败！");
+        }
         return j;
     }
 }
